Modul 3 - Pemrosesan Citra/requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
Pillow>=8.0.0
imutils>=0.5.4
